The Mute List 2.2.3
Removed FediBird after being requested to do so by @mima@makai.chaotic.ninja and also made aware of the fact that they were in the CSV and Plaintext but not mentioned in the Markdown file. This was an error on my behalf and I take full responsibility. I apologize to the Admin of Fedibird and hope this will not have longer lasting effects on their project and their instance.